Fix issues when bill includes messages or there's an acceptable failed check
The regex is modified to prevent data embedded in message callouts from
matching the regex that's crafted for the main body of the bill.

Added click to handle arg parsing, and flags for skipping checks and
printing with more verbosity to diagnose check failures.

The file path argument supports shell-expansion, such as `*`.